ai_system/requirements.txt
fastapi>=0.85.0
uvicorn>=0.17.0
sqlalchemy>=1.4
celery>=5.2
chromadb>=0.4.0
insightface>=0.7.3
opencv-python>=4.8.0
onnxruntime>=1.15.0
numpy>=1.23.0
jinja2>=3.1.2
pytest>=7.0
redis>=4.5.0
python-multipart>=0.0.6
httpx>=0.24.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
Pillow>=9.5.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2django>=6.0.3
djangorestframework>=3.17.1
django-environ>=0.13.0
